Complete 'Assignment 25'
main.js: Defined function 'assign25' which removes the KVP 'Genre' from the defined dictionary.
index.html: Added an event handler to a p element which attempts to output the deleted KVP by running JS function 'assign25' and storing the output in id 'Deleted' and displaying the result in the browser. This will return 'undefined'